Introduce a new format for level collections.
The new level file format is a XML-to-JSON conversion of SLC files that
are available in the zip file located at this URL:
http://www.sourcecode.se/sokoban/download/Levels. Also in this commit is
a dirty, but effective function that does the conversion via some search
and replace statements. According to the XML schema definition found in
http://www.sourcecode.se/sokoban/SokobanLev.xsd, there is only one
LevelCollection element. I believe that additional level of
"indentation" isn't absolutely necessary to describe the data, but it is
nevertheless retained in the conversion.